requirements.txt
aniso8601==9.0.1
asgiref==3.6.0
attrs==22.2.0
bcrypt==4.0.1
beautifulsoup4==4.11.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.1
defusedxml==0.7.1
distro==1.8.0
Django==3.2
django-cors-headers==3.14.0
django-crispy-forms==2.0
django-filter==22.1
django-safedelete==1.3.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
docker==6.0.1
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
drf-yasg==1.21.5
Faker==12.0.1
graphene==3.2.2
graphene-django==3.0.0
graphql-core==3.2.3
graphql-relay==3.2.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==3.2.0
Markdown==3.4.1
markdownify==0.11.6
MarkupSafe==2.1.2
mixer==7.2.2
oauthlib==3.2.2
packaging==23.0
paramiko==3.1.0
postgres==4.0
promise==2.3
psycopg2==2.9.6
psycopg2-binary==2.9.6
psycopg2-pool==1.1
pycparser==2.21
PyJWT==2.6.0
PyNaCl==1.5.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==0.21.1
python3-openid==3.2.0
pytz==2022.7.1
PyYAML==5.4.1
requests==2.28.2
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.4.0
soupsieve==2.4
sqlparse==0.4.3
text-unidecode==1.3
texttable==1.6.7
uritemplate==4.1.1
urllib3==1.26.15
websocket-client==0.59.0
